fix: track card rerenders on checkbox change
