+++ .gitignore
@@ -30,3 +30,4 @@ dist-ssr
 /blob-report/
 /playwright/.cache/
 test-results.json
+.turbo
