+++ src/features/tracks/components/select-all-tracks-button.tsx
@@ -6,12 +6,13 @@ import useTracksStore from '@/features/tracks/store/use-tracks-store.tsx';
 
 const SelectAllTracksButton = () => {
   const { tracks } = useTracksQuery();
-  const {
-    isSelectMode,
-    toggleSelectMode,
-    selectedTracksIds,
-    setSelectedTracksIds,
-  } = useTracksStore();
+
+  const isSelectMode = useTracksStore((state) => state.isSelectMode);
+  const selectedTracksIds = useTracksStore((state) => state.selectedTracksIds);
+  const toggleSelectMode = useTracksStore((state) => state.toggleSelectMode);
+  const setSelectedTracksIds = useTracksStore(
+    (state) => state.setSelectedTracksIds
+  );
 
   const handleSelectAll = () => {
     const allTracksIds = tracks.map((track) => track.id);
