+++ tests/integration/use-tracks.query.test.tsx
@@ -3,7 +3,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { vi, describe, it, expect, beforeEach, afterEach } from 'vitest';
 import React from 'react';
 
-import useTracksQuery from '../../src/features/tracks/hooks/use-tracks-query';
+import useTracksQuery from '../../src/features/tracks/hooks/queries/use-tracks-query';
 
 vi.mock('@/hooks/use-debounce.ts', () => ({
   default: vi.fn((value) => value),
@@ -26,12 +26,12 @@ vi.mock('@/features/tracks/api/tracks.api', () => ({
   getTracks: vi.fn(),
 }));
 
-import useTracksSearch from '../../src/features/tracks/hooks/use-tracks-search.ts';
-import { usePagination } from '../../src/hooks/use-pagination.ts';
+import useTracksSearch from '../../src/features/tracks/hooks/use-tracks-search';
+import { usePagination } from '../../src/hooks/use-pagination';
 import {
   filterTracks,
   serialize,
-} from '../../src/features/tracks/lib/utils.ts';
+} from '../../src/features/tracks/lib/utils';
 import { getTracks } from '../../src/features/tracks/api/tracks.api';
 
 const mockUseTracksSearch = vi.mocked(useTracksSearch);
